+++ src/lib/pipeline/M1PortAllocator.cc
@@ -62,7 +62,7 @@ void M1PortAllocator::issued(uint16_t port) {
 void M1PortAllocator::deallocate(uint16_t port) { issued(port); }
 
 void M1PortAllocator::setRSSizeGetter(
-    std::function<void(std::vector<uint64_t>&)> rsSizes) {
+    std::function<void(std::vector<uint32_t>&)> rsSizes) {
   rsSizes_ = rsSizes;
 }
